+++ libraries/inspector_frame/src/main/java/androidx/media3/inspector/frame/FrameExtractorInternal.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package androidx.media3.inspector;
+package androidx.media3.inspector.frame;
 
 import static android.graphics.Bitmap.Config.ARGB_8888;
 import static android.graphics.Bitmap.Config.RGBA_1010102;